+++ mobile/lib/features/trade/trade_bottom_sheet_tab.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:get_10101/common/amount_text.dart';
 import 'package:get_10101/common/amount_text_field.dart';
@@ -22,6 +20,7 @@ import 'package:get_10101/features/trade/trade_dialog.dart';
 import 'package:get_10101/features/trade/trade_theme.dart';
 import 'package:get_10101/features/trade/trade_value_change_notifier.dart';
 import 'package:go_router/go_router.dart';
+import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 
 const contractSymbol = ContractSymbol.btcusd;
@@ -57,13 +56,6 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
     super.initState();
   }
 
-  Future<rust.TradeConstraints> _getTradeConstraints() async {
-    var completer = Completer<rust.TradeConstraints>();
-    var channelTradeConstraints = rust.api.channelTradeConstraints();
-    completer.complete(channelTradeConstraints);
-    return completer.future;
-  }
-
   @override
   void dispose() {
     marginController.dispose();
@@ -81,6 +73,9 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
     String label = direction == Direction.long ? "Buy" : "Sell";
     Color color = direction == Direction.long ? tradeTheme.buy : tradeTheme.sell;
 
+    final channelInfoService = lspChangeNotifier.channelInfoService;
+    final channelTradeConstraints = channelInfoService.getTradeConstraints();
+
     return Form(
       key: _formKey,
       child: Column(
@@ -88,52 +83,14 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
         crossAxisAlignment: CrossAxisAlignment.center,
         mainAxisSize: MainAxisSize.min,
         children: [
-          FutureBuilder<rust.TradeConstraints>(
-            future: _getTradeConstraints(), // a previously-obtained Future<String> or null
-            builder: (BuildContext context, AsyncSnapshot<rust.TradeConstraints> snapshot) {
-              List<Widget> children;
-
-              final channelInfoService = lspChangeNotifier.channelInfoService;
-
-              if (snapshot.hasData) {
-                var tradeConstraints = snapshot.data!;
-
-                children = <Widget>[
-                  buildChildren(direction, tradeConstraints, context, channelInfoService, _formKey),
-                ];
-              } else if (snapshot.hasError) {
-                children = <Widget>[
-                  const Icon(
-                    Icons.error_outline,
-                    color: Colors.red,
-                    size: 60,
-                  ),
-                  Padding(
-                    padding: const EdgeInsets.only(top: 16),
-                    child:
-                        Text('Error: Could not load confirmation screen due to ${snapshot.error}'),
-                  ),
-                ];
-              } else {
-                children = const <Widget>[
-                  SizedBox(
-                    width: 60,
-                    height: 60,
-                    child: CircularProgressIndicator(),
-                  ),
-                  Padding(
-                    padding: EdgeInsets.only(top: 16),
-                    child: Text('Loading confirmation screen...'),
-                  ),
-                ];
-              }
-              return Center(
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: children,
-                ),
-              );
-            },
+          Center(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                buildChildren(
+                    direction, channelTradeConstraints, context, channelInfoService, _formKey)
+              ],
+            ),
           ),
           Column(
             crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -141,9 +98,10 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
               ElevatedButton(
                   key: widget.buttonKey,
                   onPressed: () {
-                    if (_formKey.currentState!.validate()) {
-                      TradeValues tradeValues =
-                          context.read<TradeValuesChangeNotifier>().fromDirection(direction);
+                    TradeValues tradeValues =
+                        context.read<TradeValuesChangeNotifier>().fromDirection(direction);
+                    if (_formKey.currentState!.validate() &&
+                        channelTradeConstraints.minMargin <= (tradeValues.margin?.sats ?? 0)) {
                       final submitOrderChangeNotifier = context.read<SubmitOrderChangeNotifier>();
                       tradeBottomSheetConfirmation(
                           context: context,
@@ -209,6 +167,8 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
           "\nWith your current balance, the maximum you can trade is ${formatUsd(Usd(maxQuantity.toInt()))}";
     }
 
+    var amountFormatter = NumberFormat.compact(locale: "en_UK");
+
     return Wrap(
       runSpacing: 12,
       children: [
@@ -216,7 +176,7 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
           padding: const EdgeInsets.only(bottom: 10),
           child: Row(
             children: [
-              const Flexible(child: Text("Usable Balance:")),
+              const Flexible(child: Text("Balance:")),
               const SizedBox(width: 5),
               Flexible(child: AmountText(amount: Amount(usableBalance))),
               const SizedBox(
@@ -319,6 +279,9 @@ class _TradeBottomSheetTabState extends State<TradeBottomSheetTab> {
                       return AmountTextField(
                         value: margin,
                         label: "Margin (sats)",
+                        error: channelTradeConstraints.minMargin > margin.sats
+                            ? "Min margin is ${amountFormatter.format(channelTradeConstraints.minMargin)} sats"
+                            : null,
                         suffixIcon: showCapacityInfo
                             ? ModalBottomSheetInfo(
                                 closeButtonText: "Back to order",